refactor(#251): MemberList 리팩토링 및 test 작성
