+++ src/widgets/team/GoalList/GoalList.tsx
@@ -7,7 +7,7 @@ import { useTeamId } from "@/features/team";
 import { useInfiniteScroll } from "@/shared/hooks/useInfiniteScroll";
 import { Icon } from "@/shared/ui/Icon";
 import { Order } from "@/shared/ui/Order";
-import { GoalProgressCard } from "@/widgets/team/GoalProgressCard";
+import { GoalProgressCard } from "@/widgets/team/GoalProgressCard/GoalProgressCard";
 
 const sortOptions = ["최신순", "오래된순"];
